Refatorando código e ajustando Toast
