+++ app/src/main/AndroidManifest.xml
@@ -18,10 +18,8 @@
         <activity android:name=".PendenciasActivity"></activity>
         <activity android:name=".ListaAlunosSQLiteActivity" />
         <activity android:name=".MainSQLiteActivity" />
-        <activity android:name=".NewLoginActivity"></activity>
         <activity android:name=".CadastroActivity"></activity>
-        <activity android:name=".PrincipalActivity"></activity>
-        <activity android:name=".NovoCadastro"></activity>
+        <activity android:name=".LoginActivity"></activity>
         <activity android:name=".DeBug"></activity>
         <activity android:name=".SplashActivity">
             <intent-filter>
@@ -30,7 +28,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LoginActivity" />
+        <activity android:name=".MainActivity" />
 
         <meta-data
             android:name="preloaded_fonts"
