+++ app/src/main/java/br/edu/uniritter/mobile/grupo_2_projeto_final/SplashActivity.java
@@ -3,7 +3,6 @@ package br.edu.uniritter.mobile.grupo_2_projeto_final;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.view.View;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -20,7 +19,7 @@ public class SplashActivity extends AppCompatActivity {
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
-                startActivity(new Intent(getBaseContext(), LoginActivity.class));
+                startActivity(new Intent(getBaseContext(), MainActivity.class));
                 finish();
             }
         }, SPLASH_TIME_OUT);
