+++ app/src/main/res/layout/activity_primeira.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/backActivities"
-    tools:context=".LoginActivity">
+    tools:context=".MainActivity">
 
     <Button
         android:id="@+id/btCadastrar"
